+++ src/typed_instrs.rs
@@ -2,7 +2,7 @@ use crate::elem::ElemSymbol;
 use crate::elem_type::{ElemType, StackType};
 use crate::stack::Stack;
 use crate::restack::Restack;
-use crate::types_scratch::ElemsPopError;
+use crate::elems::ElemsPopError;
 use crate::typed_instruction::{IsStackInstruction, StackInstructionError};
 use crate::typed_instr::Instr;
