+++ src/typed_instructions.rs
@@ -1,7 +1,10 @@
 use crate::elem::{Elem, ElemSymbol};
 use crate::an_elem::AnElem;
 use crate::types::{Empty, Nil};
-use crate::types_scratch::{AllElems, Singleton, Cons, ReturnSingleton, ConsOut, Or, ReturnOr, IsList};
+use crate::elems_singleton::Singleton;
+use crate::elems_or::Or;
+use crate::elems_all::AllElems;
+use crate::types_scratch::{Cons, ReturnSingleton, ConsOut, ReturnOr, IsList};
 use crate::untyped_instruction::Instruction;
 use crate::typed_instruction::{IsInstructionT, StackInstructionError};
 
@@ -377,7 +380,7 @@ impl IsInstructionT for ToJson {
     fn run(&self, x: &Self::IO) -> Result<(), Self::Error> {
         let returning = x.clone().hd().returning;
         let y = &x.clone().tl().hd();
-        let array = y.all_elems_untyped();
+        let array = y.untyped();
         let z = array[0].clone();
         returning.returning(serde_json::to_value(z.clone())
                             .map_err(move |e| ToJsonError {
@@ -617,7 +620,7 @@ impl IsInstructionT for CheckLe {
     fn run(&self, x: &Self::IO) -> Result<(), Self::Error> {
         let returning = x.clone().hd().returning;
         let y = &x.clone().tl().hd();
-        let array = y.all_elems_untyped();
+        let array = y.untyped();
         let lhs = array[0].clone();
         let rhs = array[1].clone();
         let cmp_result = lhs.partial_cmp(&rhs)
@@ -664,7 +667,7 @@ impl IsInstructionT for CheckLt {
     fn run(&self, x: &Self::IO) -> Result<(), Self::Error> {
         let returning = x.clone().hd().returning;
         let y = &x.clone().tl().hd();
-        let array = y.all_elems_untyped();
+        let array = y.untyped();
         let lhs = array[0].clone();
         let rhs = array[1].clone();
         let cmp_result = lhs.partial_cmp(&rhs)
@@ -711,7 +714,7 @@ impl IsInstructionT for CheckEq {
     fn run(&self, x: &Self::IO) -> Result<(), Self::Error> {
         let returning = x.clone().hd().returning;
         let y = &x.clone().tl().hd();
-        let array = y.all_elems_untyped();
+        let array = y.untyped();
         let lhs = array[0].clone();
         let rhs = array[1].clone();
         let cmp_result = lhs.partial_cmp(&rhs)
